requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
